Day06:User CRUD Test

+++ src/main/java/com/ezen/antpeople/repository/UserRepository.java
@@ -3,10 +3,9 @@ package com.ezen.antpeople.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ezen.antpeople.dto.UserDTO;
 import com.ezen.antpeople.entity.UserEntity;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long>{
+public interface UserRepository extends JpaRepository<UserEntity, Integer>{
 	UserEntity findByEmail(String email);
 }
